did stuff

+++ src/main/java/client/wwwings/Player.java
@@ -4,11 +4,11 @@ import java.util.UUID;
 public class Player {
     private String uuid;
     private String name;
-    private int[] position;
-    public Player(String uuid, String name, int x, int y, int z) {
+    private double[] position;
+    public Player(String uuid, String name, double x, double y, double z) {
         this.name = name;
         this.uuid = uuid;
-        this.position = new int[]{x, y, z};
+        this.position = new double[]{x, y, z};
     }
 
     public String getUuid() {
